requirements.txt
#  dify_plugin
dify_plugin~=0.2.2

# for doc
pypandoc-binary~=1.15.0
beautifulsoup4 ~=4.12.2
python-docx~=1.1.2 
markdown~=3.8.0

python-pptx~=1.0.2
